Move mod/filerm to src/Module/FilerM
